Un-refactor
